feat(agents-gallery): moves agents list to another file
